# docs/swagger.json
{
    "schemes": [
        "https",
        "http"
    ],
    "swagger": "2.0",
    "info": {
        "description": "Meta App Service Indexer Service API, provides MetaApp query and deploy functionality",
        "title": "Meta App Service Indexer API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "API Support",
            "url": "http://www.swagger.io/support",
            "email": "support@swagger.io"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0"
    },
    "host": "localhost:7333",
    "basePath": "/",
    "paths": {
        "/api/v1/config": {
            "get": {
                "description": "获取前端需要的配置信息，如 Metafs Domain",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Indexer Status"
                ],
                "summary": "获取配置信息",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/meta-app-service_controller_respond.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/meta-app-service_controller_respond.ConfigResponse"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/meta-app-service_controller_respond.Response"
                        }
                    }
                }
            }
        },
        "/api/v1/deploy-queue": {
            "get": {
                "description": "获取部署队列列表，按时间戳倒序排列，支持分页",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Deploy Queue"
                ],
                "summary": "获取部署队列列表",
                "parameters": [
                    {
                        "type": "integer",
                        "default": 0,
                        "description": "游标（从 0 开始）",
                        "name": "cursor",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "default": 20,
                        "description": "每页大小",
                        "name": "size",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/meta-app-service_controller_respond.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/meta-app-service_controller_respond.DeployQueueListResponse"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/meta-app-service_controller_respond.Response"
                        }
                    }
                }
            }
        },
        "/api/v1/metaapps": {
            "get": {
                "description": "获取所有 MetaApp 列表，按时间倒序排列，支持分页",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "MetaApp"
                ],
                "summary": "获取 MetaApp 列表",
                "parameters": [
                    {
                        "type": "integer",
                        "default": 0,
                        "description": "游标（从 0 开始）",
                        "name": "cursor",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "default": 20,
                        "description": "每页大小",
                        "name": "size",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/meta-app-service_controller_respond.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/meta-app-service_controller_respond.MetaAppListResponse"
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            }
        },
        "/api/v1/metaapps/creator/{metaId}": {
            "get": {
                "description": "根据创建者 MetaID 获取 MetaApp 列表，包括部署情况，按时间倒序排列，支持分页",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "MetaApp"
                ],
                "summary": "根据 MetaID 获取 MetaApp 列表",
                "parameters": [
                    {
                        "type": "string",
                        "description": "创建者 MetaID",
                        "name": "metaId",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "default": 0,
                        "description": "游标（从 0 开始）",
                        "name": "cursor",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "default": 20,
                        "description": "每页大小",
                        "name": "size",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/meta-app-service_controller_respond.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/meta-app-service_controller_respond.MetaAppListResponse"
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            }
        },
        "/api/v1/metaapps/first/{firstPinId}": {
            "get": {
                "description": "根据 FirstPinID 获取最新的 MetaApp 详细信息，包括部署情况",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "MetaApp"
                ],
                "summary": "根据 FirstPinID 获取最新的 MetaApp 详情",
                "parameters": [
                    {
                        "type": "string",
                        "description": "MetaApp FirstPinID",
                        "name": "firstPinId",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/meta-app-service_controller_respond.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/meta-app-service_controller_respond.MetaAppResponse"
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            }
        },
        "/api/v1/metaapps/first/{firstPinId}/download": {
            "get": {
                "description": "根据 FirstPinID 压缩对应的部署文件夹并下载为 zip 文件",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/zip"
                ],
                "tags": [
                    "MetaApp"
                ],
                "summary": "下载 MetaApp 部署文件为 zip",
                "parameters": [
                    {
                        "type": "string",
                        "description": "MetaApp FirstPinID",
                        "name": "firstPinId",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "zip file",
                        "schema": {
                            "type": "file"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/meta-app-service_controller_respond.Response"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/meta-app-service_controller_respond.Response"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/meta-app-service_controller_respond.Response"
                        }
                    }
                }
            }
        },
        "/api/v1/metaapps/first/{firstPinId}/history": {
            "get": {
                "description": "根据 FirstPinID 获取 MetaApp 的所有历史版本列表",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "MetaApp"
                ],
                "summary": "根据 FirstPinID 获取 MetaApp 历史版本列表",
                "parameters": [
                    {
                        "type": "string",
                        "description": "MetaApp FirstPinID",
                        "name": "firstPinId",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/meta-app-service_controller_respond.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/meta-app-service_controller_respond.MetaAppHistoryResponse"
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            }
        },
        "/api/v1/metaapps/{pinId}": {
            "get": {
                "description": "根据 PinID 获取 MetaApp 详细信息，包括部署情况",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "MetaApp"
                ],
                "summary": "根据 PinID 获取 MetaApp 详情",
                "parameters": [
                    {
                        "type": "string",
                        "description": "MetaApp PinID",
                        "name": "pinId",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/meta-app-service_controller_respond.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/meta-app-service_service_indexer_service.MetaAppWithDeploy"
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            }
        },
        "/api/v1/metaapps/{pinId}/redeploy": {
            "post": {
                "description": "根据 PinID 重新将 MetaApp 加入部署队列",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "MetaApp"
                ],
                "summary": "重新部署 MetaApp",
                "parameters": [
                    {
                        "type": "string",
                        "description": "MetaApp PinID",
                        "name": "pinId",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/meta-app-service_controller_respond.Response"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/meta-app-service_controller_respond.Response"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/meta-app-service_controller_respond.Response"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/meta-app-service_controller_respond.Response"
                        }
                    }
                }
            }
        },
        "/api/v1/stats": {
            "get": {
                "description": "获取索引器统计信息（当前已同步的 MetaApp 总数）",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Indexer Status"
                ],
                "summary": "获取统计信息",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/meta-app-service_controller_respond.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/meta-app-service_controller_respond.IndexerStatsResponse"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/meta-app-service_controller_respond.Response"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "meta-app-service_controller_respond.ConfigResponse": {
            "type": "object",
            "properties": {
                "metafs_domain": {
                    "type": "string",
                    "example": "http://localhost:7281"
                }
            }
        },
        "meta-app-service_controller_respond.DeployQueueListResponse": {
            "type": "object",
            "properties": {
                "has_more": {
                    "type": "boolean",
                    "example": true
                },
                "next_cursor": {
                    "type": "integer",
                    "example": 100
                },
                "queues": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/meta-app-service_controller_respond.DeployQueueResponse"
                    }
                }
            }
        },
...
        "meta-app-service_controller_respond.MetaAppHistoryResponse": {
            "type": "object",
            "properties": {
                "history": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/meta-app-service_controller_respond.MetaAppResponse"
                    }
                }
            }
        },
        "meta-app-service_controller_respond.MetaAppListResponse": {
            "type": "object",
            "properties": {
                "apps": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/meta-app-service_controller_respond.MetaAppResponse"
                    }
                },
                "has_more": {
                    "type": "boolean",
                    "example": true
                },
                "next_cursor": {
                    "type": "integer",
                    "example": 100
                }
            }
        },
...
        "meta-app-service_controller_respond.Response": {
            "description": "Unified API response structure",
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer",
                    "example": 0
                },
                "data": {},
                "message": {
                    "type": "string",
                    "example": "success"
                },
                "processingTime": {
                    "type": "integer",
                    "example": 123
                }
            }
        },
...
    }
}